adding fyrst testing class
